+++ src/OpenSSL/debug/__main__.py
@@ -1,26 +1,26 @@
 from __future__ import print_function
 
+import ssl
 import sys
 
 import OpenSSL.SSL
 import cryptography
 
-from . import version
+from .. import version
 
 
-if __name__ != "__main__":
-    raise ImportError("This module can't be imported.")
-
-wot = """\
+_env_info = u"""\
 pyOpenSSL: {pyopenssl}
 cryptography: {cryptography}
-OpenSSL: {openssl}
-Python: {python}
+cryptography's OpenSSL: {crypto_openssl}
+Pythons's OpenSSL: {python_openssl}
+Python executable: {python}
 Python version: {python_version}
 Platform: {platform}
 sys.path: {sys_path}""".format(
     pyopenssl=version.__version__,
-    openssl=OpenSSL.SSL.SSLeay_version(OpenSSL.SSL.SSLEAY_VERSION),
+    crypto_openssl=OpenSSL.SSL.SSLeay_version(OpenSSL.SSL.SSLEAY_VERSION),
+    python_openssl=ssl.OPENSSL_VERSION,
     cryptography=cryptography.__version__,
     python=sys.executable,
     python_version=sys.version,
@@ -28,4 +28,6 @@ sys.path: {sys_path}""".format(
     sys_path=sys.path,
 )
 
-print(wot)
+
+if __name__ == "__main__":
+    print(_env_info)