+++ tests/test_util.py
@@ -1,6 +1,6 @@
-from OpenSSL._util import exception_from_error_queue, lib
-
 import pytest
+
+from OpenSSL._util import exception_from_error_queue, lib
 
 
 class TestErrors(object):
@@ -9,7 +9,7 @@ class TestErrors(object):
     """
     def test_exception_from_error_queue_nonexistent_reason(self):
         """
-        :py:func:`exception_from_error_queue` raises ``ValueError`` when it
+        :func:`exception_from_error_queue` raises ``ValueError`` when it
         encounters an OpenSSL error code which does not have a reason string.
         """
         lib.ERR_put_error(lib.ERR_LIB_EVP, 0, 1112, b"", 10)