+++ CHANGELOG.rst
@@ -17,8 +17,9 @@ Backward-incompatible changes:
   `#630 <https://github.com/pyca/pyopenssl/pull/630>`_
 - Fixed a bug with ``ASN1_TIME`` casting in ``X509.set_notBefore``,
   ``X509.set_notAfter``, ``Revoked.set_rev_date``, ``Revoked.set_nextUpdate``,
-  and ``Revoked.set_lastUpdate``. You must now pass times solely in the form
-  ``YYYYMMDDhhmmssZ`` `#612 <https://github.com/pyca/pyopenssl/pull/612>`_
+  and ``Revoked.set_lastUpdate``. You must now pass times in the form
+  ``YYYYMMDDhhmmssZ``. ``YYYYMMDDhhmmss+hhmm`` and ``YYYYMMDDhhmmss-hhmm``
+  will no longer work. `#612 <https://github.com/pyca/pyopenssl/pull/612>`_
 
 
 Deprecations: