+++ CHANGELOG.rst
@@ -15,6 +15,11 @@ Backward-incompatible changes:
 - Removed the deprecated ``OpenSSL.rand.egd`` function.
   Applications should prefer ``os.urandom()`` for random number generation.
   `#630 <https://github.com/pyca/pyopenssl/pull/630>`_
+- Fixed a bug with ``ASN1_TIME`` casting in ``X509.set_notBefore``,
+  ``X509.set_notAfter``, ``Revoked.set_rev_date``, ``Revoked.set_nextUpdate``,
+  and ``Revoked.set_lastUpdate``. You must now pass times in the form
+  ``YYYYMMDDhhmmssZ``. ``YYYYMMDDhhmmss+hhmm`` and ``YYYYMMDDhhmmss-hhmm``
+  will no longer work. `#612 <https://github.com/pyca/pyopenssl/pull/612>`_
 
 
 Deprecations: